requirements.txt
# Python Tracker Requirements
# Install with: pip install -r requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0
pywin32>=306
psutil>=5.9.0
pynput>=1.7.6
